Question_1/event_platform_backend/event_platform/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg[binary]==3.3.3
Pillow==12.1.1
python-decouple==3.8
django-cors-headers==4.3.1